Ajout de différents filtres sur la page de recherche

+++ topoguide/itineraires/models.py
@@ -24,7 +24,7 @@ class Itineraire(models.Model):
     CHOIX_DIF = ((1,'1'),(2,'2'),(3,'3'), (4,'4'), (5,'5')) # Liste d'entier avec choix pour difficulté
     difficulte = models.IntegerField('Difficulté (de 1 à 5)', default=1,choices=CHOIX_DIF)
 
-    
+
     def __str__(self):
         return self.titre
     
@@ -38,7 +38,7 @@ class Sortie(models.Model):
     utilisateur = models.ForeignKey(User, on_delete=models.CASCADE) #Référence à enregistrements d'autres tables avec le type ForeignKey
     itineraire = models.ForeignKey(Itineraire, on_delete=models.CASCADE)
     date_sortie = models.DateField('Date de la sortie')
-    duree_reelle = models.FloatField('Durée réelle (en heure)')
+    duree_reelle = models.IntegerField('Durée réelle (en heure)')
     CHOIX_EXP = (('Tous débutants','Tous débutants'),('Tous expérimentés','Tous expérimentés'),('Mixte','Mixte')) # Liste de caractères avec choix pour expérience
     nombre_personne = models.FloatField('Nombre de personnes ayant réalisé la sortie')
     experience = models.CharField('Expérience du groupe', max_length= 20,choices=CHOIX_EXP)
